+++ gatsby-transformer-kickstartds-netlify-cms/CHANGELOG.md
@@ -1,3 +1,27 @@
+# v1.1.4 (Tue Jul 13 2021)
+
+#### ⚠️ Pushed to `master`
+
+- feat: load tokens inside Netlify CMS preview ([@julrich](https://github.com/julrich))
+
+#### Authors: 1
+
+- Jonas Ulrich ([@julrich](https://github.com/julrich))
+
+---
+
+# v1.1.3 (Mon Jul 12 2021)
+
+#### ⚠️ Pushed to `master`
+
+- fix: quote element hashing, add images ([@julrich](https://github.com/julrich))
+
+#### Authors: 1
+
+- Jonas Ulrich ([@julrich](https://github.com/julrich))
+
+---
+
 # v1.1.2 (Sun Jul 11 2021)
 
 #### ⚠️ Pushed to `master`